refactor: (#56) SignUp -> Signup으로 이름 변경 및 Route 적용
