+++ src/pages/index.ts
@@ -1,5 +1,5 @@
 export * from './LoginPage';
 export * from './MainPage';
 export * from './RoomPage';
-export * from './SignUpPage';
+export * from './SignupPage';
 export * from './WelcomePage';
